feat(client): edit client
